use custom mp3player included in wiiradio in Makefile

+++ libmp3player_wiiradio/mp3player_wiiradio.h
@@ -1,7 +1,7 @@
 #ifndef __MP3PLAYER_H__
 #define __MP3PLAYER_H__
 
-#include <mad/mad.h>
+#include <mad.h>
 #include <gctypes.h>
 
 #ifdef __cplusplus
@@ -13,7 +13,7 @@ void MP3Player_Stop();
 BOOL MP3Player_IsPlaying();
 void MP3Player_Volume(u32 volume);
 //s32 MP3Player_PlayBuffer(const void *buffer,s32 len,void (*filterfunc)(struct mad_stream *,struct mad_frame *));
-s32 MP3Player_PlayBuffer(const void *buffer,s32 len,void (*filterfunc)(struct mad_stream *,struct mad_frame *),void (*fftin)(u8* data,int len));
+s32 MP3Player_PlayBuffer(const void *buffer,s32 len,void (*filterfunc)(struct mad_stream *,struct mad_frame *),void (*fftin)(s16* data,int len));
 s32 MP3Player_PlayFile(void *cb_data,s32 (*reader)(void *,void *,s32),void (*filterfunc)(struct mad_stream *,struct mad_frame *),void (*fftin)(s16* data,int len));
 
 #ifdef __cplusplus
